feat: Add first_name and last_name to user creation in LoanModelTest
This commit modifies the LoanModelTest setup to include `first_name`
and `last_name` when creating a test user. This change ensures that
the user object has these attributes populated, which might be required
by other parts of the application or future tests. This avoids potential
AttributeError exceptions if `first_name` or `last_name` are accessed.