feat: Add INTERNET permission to debug AndroidManifest.xml (debug)
This commit introduces the `AndroidManifest.xml` file under the `android/app/src/debug/` directory.  This file adds the `android.permission.INTERNET` permission, which is essential for development-related functionalities such as hot reload, setting breakpoints, and communication between the Flutter tool and the running application during debugging sessions. This permission is only included in the debug build.