feat: Add launch_background.xml for splash screen customization
This commit introduces the `launch_background.xml` file to the
`drawable-v21` directory. This file defines the background of the
launch splash screen for Android API level 21 and above.  It uses a
`layer-list` to allow for customization, including setting the
background color and adding an image.  The default configuration sets
the background to the system's background color and provides a commented-out
example for adding a centered image asset (`launch_image`). This
enables developers to easily customize the app's splash screen.