feat: Add app launcher icon (xhdpi)
This commit adds the `ic_launcher.png` file to the `mipmap-xhdpi` directory.
This file serves as the application's launcher icon for devices with
xhdpi (extra-high-density) screens.  Including this icon ensures that
the app has a visual representation on the device's home screen and app
drawer for these screen densities. This improves the user experience
by providing a consistent and appropriate icon for different device
resolutions.