feat: Add initial Xcode project file for iOS Runner target
This commit introduces the 'project.pbxproj' file, which is the
core project file for the iOS Runner target in Xcode.  It defines
the project structure, build settings, targets, build phases,
file references, and other configurations necessary for building
the iOS application.

This file was automatically generated by Flutter during project
creation. It includes configurations for:

- Build phases (Sources, Frameworks, Resources, Scripts).
- Build settings for Debug, Release, and Profile configurations.
- File references for source code, assets, and configuration files.
- Target dependencies.

This commit provides the initial baseline for further iOS-specific
configurations and customizations.