feat: Add initial LaunchScreen.storyboard for iOS
This commit adds the `LaunchScreen.storyboard` file to the iOS Runner project.
The storyboard defines the app's launch screen using an `imageView` that
displays the `LaunchImage`. The image is centered within the view. This
provides a visual representation while the app loads, improving the user
experience. The background color is set to white.