feat: Add basic XCTest case for Runner app
This commit introduces a basic XCTest case within the `RunnerTests` class.
The file `sacco_mobile/ios/RunnerTests/RunnerTests.swift` has been created
containing the `RunnerTests` class, inheriting from `XCTestCase`.  It
includes a placeholder `testExample` function as a starting point for
future UI and integration tests. The comments direct the developer to
the Apple XCTest documentation. This provides a foundation for writing
tests to ensure the stability and functionality of the Flutter application.