feat: Initial plugin registrant for Linux platform
This commit adds the initial `generated_plugin_registrant.cc` file
for the Linux platform. This file is responsible for registering
Flutter plugins.  Currently, it contains an empty
`fl_register_plugins` function.  Future commits will populate this
function as plugins are added to the application. The file is
generated and should not be manually edited.