feat: Generate plugin registrant for Flutter Linux
This commit introduces the `generated_plugin_registrant.h` file,
which is responsible for registering Flutter plugins on the Linux
platform. This file includes the necessary header for Flutter Linux
(`flutter_linux/flutter_linux.h`) and declares the
`fl_register_plugins` function. This function will be used to
register all the Flutter plugins used in the application with the
Flutter engine at startup.  The file also includes header guards
to prevent multiple inclusions.  This is a standard generated file
required for Flutter Linux support.