feat: Create MyApplication class header file
This commit introduces the `my_application.h` header file. This file
declares the `MyApplication` class, a subclass of `GtkApplication`, which
serves as the foundation for the Flutter application on Linux.

The header includes necessary GTK headers and defines the GType for
`MyApplication`, facilitating its integration with the GObject type system.
It also declares the `my_application_new()` function, responsible for
creating new instances of the `MyApplication` class.

This header file is crucial for setting up the application structure
and enabling the use of Flutter within a GTK environment on Linux.