feat: Add xcworkspace file for macOS Runner project
This commit introduces the `contents.xcworkspacedata` file to the
macOS Runner project. This file defines the workspace, which includes
the `Runner.xcodeproj` project.  Adding this file is crucial for
proper Xcode workspace configuration and allows Xcode to correctly
manage the project's structure and dependencies. This ensures the
macOS app can be built and run correctly within Xcode.