feat: Initialize generated plugin registrant for Windows
This commit creates the `generated_plugin_registrant.cc` file,
which is responsible for registering Flutter plugins on Windows.
Currently, the `RegisterPlugins` function is empty. Plugins will be
registered here in future commits as they are added to the project.
This ensures the basic infrastructure for plugin support is in place.