feat: Create generated_plugin_registrant.h for Windows platform
This commit introduces the 'generated_plugin_registrant.h' file
for the Windows platform. This header file is responsible for
registering Flutter plugins within the Windows application.

The file includes the necessary header guards (`#ifndef`, `#define`,
`#endif`) to prevent multiple inclusions and declares the
`RegisterPlugins` function, which takes a `flutter::PluginRegistry`
pointer as input. This function will be used to register the
plugins with the Flutter engine.

This is a generated file and should not be manually edited.